Removes title from nav bar for UIWebView
The issue is that the title is not being updated as you move from page
to page, so it becomes invalid as soon as you follow a link in Pod.
There's a similar issue with the bookmark/trash can icon not being
updated, but I don't want to remove that functionality at this point.